// src/federated-entities/entities/enums/federative_unit.enum.ts
export enum FEDERATIVE_UNIT {
  UNIAO = 'UNIÃO',
  // STATE
  ACRE = 'AC',
  ALAGOAS = 'AL',
  AMAPA = 'AP',
  AMAZONAS = 'AM',
  BAHIA = 'BA',
  CEARA = 'CE',
  DISTRITO_FEDERAL = 'DF',
  ESPIRITO_SANTO = 'ES',
  GOIAS = 'GO',
  MARANHAO = 'MA',
  MATO_GROSSO = 'MT',
  MATO_GROSSO_DO_SUL = 'MS',
  MINAS_GERAIS = 'MG',
  PARA = 'PA',
  PARAIBA = 'PB',
  PARANA = 'PR',
  PERNAMBUCO = 'PE',
  PIAUI = 'PI',
  RIO_DE_JANEIRO = 'RJ',
  RIO_GRANDE_DO_NORTE = 'RN',
  RIO_GRANDE_DO_SUL = 'RS',
  RONDONIA = 'RO',
  RORAIMA = 'RR',
  SANTA_CATARINA = 'SC',
  SAO_PAULO = 'SP',
  SERGIPE = 'SE',
  TOCANTINS = 'TO',
  // MUNICIPALITY
  ARACAJU = 'Aracaju',
  BELEM = 'Belém',
  BELO_HORIZONTE = 'Belo Horizonte',
  BLUMENAU = 'Blumenau',
  BOA_VISTA = 'Boa Vista',
  BRASILIA = 'Brasília',
  CAMPO_GRANDE = 'Campo Grande',
  CUIABA = 'Cuiabá',
  CURITIBA = 'Curitiba',
  FLORIANOPOLIS = 'Florianópolis',
  FORTALEZA = 'Fortaleza',
  GOIANIA = 'Goiânia',
  JOAO_PESSOA = 'João Pessoa',
  MACAPA = 'Macapá',
  MACEIO = 'Maceió',
  MANAUS = 'Manaus',
  NATAL = 'Natal',
  PALMAS = 'Palmas',
  PORTO_ALEGRE = 'Porto Alegre',
  PORTO_VELHO = 'Porto Velho',
  RECIFE = 'Recife',
  COUNTY_RIO_DE_JANEIRO = 'Rio de Janeiro',
  SALVADOR = 'Salvador',
  SAO_LUIS = 'São Luís',
  COUNTY_SAO_PAULO = 'São Paulo',
  TERESINA = 'Teresina',
  VITORIA = 'Vitória',
}